requirements.txt
newspaper3k
news-please
